Reduced debug log verbosity and fixed compilation warnings

+++ lib/parser.c
@@ -44,7 +44,7 @@ parser_feed(struct parser *p, const uint8_t c) {
         const int when = state[i].when;
         if (state[i].when <= 0xFF) {
             matched = (c == when);
-        } else if(state[i].when == ANY) {
+        } else if(((unsigned) state[i].when) == ANY) {
             matched = true;
         } else if(state[i].when > 0xFF) {
             matched = (type & when);
